pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=6.2.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.0",
    "isort>=7.0.0",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "smolagents[mcp,openai,toolkit]>=1.23.0",
]
